requirements.txt
pydantic
jinja2
typer
langgraph
matplotlib
pandas
Numpy
python-dotenv
langgraph-cli
langgraph-api
openai
mcp>=0.1.0
faiss-cpu 
sentence-transformers